apis_services/mqToMastodon/requirements.txt
# requirements file for the mastodon weather poster
Mastodon.py
paho-mqtt
